requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
google-api-core>=2.15.0
google-auth>=2.28.0
faiss-cpu>=1.8.0
google-generativeai>=0.8.0
yfinance>=0.2.40
alpha_vantage>=2.3.1
streamlit>=1.40.0
plotly>=5.24.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.32.0
beautifulsoup4>=4.12.0
tiktoken>=0.7.0
sentence-transformers>=3.0.0
langchain-huggingface>=0.1.0
mcp>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
PyYAML>=6.0.1
tenacity>=8.5.0
